moved assets to new position
