+++ client/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import store from './old-ui-src/store';
+import store from './app/store';
 
 import { Provider } from 'react-redux';
