+++ client/src/old-ui-src/components/layout/dashboard/Header.jsx
@@ -16,7 +16,7 @@ import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
 import InboxIcon from "@material-ui/icons/MoveToInbox";
 import MailIcon from "@material-ui/icons/Mail";
-import logo from "../../../assets/images/logo.png";
+import logo from "../../../../assets/images/logo.png";
 
 const drawerWidth = 240;
