+++ client/src/old-ui-src/components/layout/landing-pages/Features.jsx
@@ -1,4 +1,4 @@
-import "../../../assets/css/features.css";
+import "../../../../assets/css/features.css";
 import { Box, Grid, Paper, Button, useMediaQuery } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import Divider from "../../utils/Divider";
