+++ client/src/old-ui-src/components/layout/landing-pages/LandingHeader.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, Fragment } from "react";
-import logo from "../../../assets/images/logo.png";
+import logo from "../../../../assets/images/logo.png";
 import AuthButtons from "./AuthButtons";
 import HeaderLink from "../../utils/StyledHeaderLink";
 import { headerLinks } from "./data/landing-page";
